+++ plugin/src/integrationTest/kotlin/com/vanniktech/maven/publish/ProjectSpecRunner.kt
@@ -171,6 +171,7 @@ private fun writeSettingFile(path: Path) {
             mavenLocal()
             mavenCentral()
             google()
+            gradlePluginPortal()
         }
     }